Add files via upload
Any signal undergoing multi-path propagation or changing direction at multiple instances during its whole journey can be genarally treated as random walks. The simple examples are ultrasound signal, communication wireless signals , light wave etc. This is more often seen in the backscattered signal. Backscattering is the reflection of a wave (such as a radio wave, light wave, or acoustic wave) off a surface or interface, in the direction opposite to the direction of the incoming wave.